# Added support for custom width providers