@mgeisler 2018-04-06
Hey @Moxinilian, thanks for the PR. It's not  been forgotten, I've just been busy lately...

I'm traveling this weekend, so it won't be before next week that I get proper time to look at this.

Just some short remarks: I love that you could do this via a generic function parameter --- zero-cost abstractions are awesome, as you say :-)

About measuring the characters: would it perhaps be better to measure words at a time?

With combining characters, I guess you don't really know the true displayed width of a word until you typeset the entire thing. Take emoticons as an extreme example since I'm typing this on my phone: I recently learned that a symbol like 👩‍👩‍👧‍👦 is created at rendering time --- it's really multiple Unicode characters with zero-width-joiners between them, something like 👧👧👨👨 next to each other. See https://emojipedia.org/emoji-zwj-sequences/.

So if you measure the width of this `char` by `char`, then you get the wrong result. I think other combining characters will have the same problem.

Finally, you're definitely right that the crate should be refactored by now... It was my first larger piece of Rust code, so I'm happy to learn how to structure it better!