@Moxinilian 2018-04-07
I'm going to revert the changes and try to implement that instead, as there would be no point in merging this now that we're going down the word route.