@Moxinilian 2018-04-07
I'm going to read this document to verify easy solutions have not already been found.
http://unicode.org/reports/tr14/