@robinkrahl 2020-10-18
I’m also interested in this feature to be able to use `textwrap` with PDF documents (in [`genpdf`](https://lib.rs/crates/genpdf)).  What is the current status of this PR?  Is there anything I can do to help?