@mgeisler 2020-10-18
Hi @robinkrahl, I think this PR has stalled, but I would still like to see this kind of feature implemented. @Moxinilian, if you're still interested in this, then please let us know!

I recently rewrote some of the inner parts of textwrap (#213), so I think this PR should be closed and then a new one could be opened to fit with the new API. Ideally, this change should be invisible at the surface-level of the API, so that one can continue to write

```
let lines = wrap("some text", 80);
```

Perhaps the `WrapOptions` trait could gain a `width(&str) -> T` method, where `T` can be `usize`, `float`, or even `u16` as needed (I mention `u16` since the `terminal_size` crate [measures the width in  `u16`](http://eminence.github.io/terminal-size/doc/terminal_size/struct.Width.html), which makes me think that it's completely overkill to use `usize` all over).

Let me close this PR for now and then we can perhaps discuss more in #126.